Split libp2p_helper main to a few files
